Flytt dokumentÅrsakTyper ut i egen fil
