+++ src/frontend/sider/Fagsak/Dokumentutsending/DokumentutsendingSkjema.tsx
@@ -10,7 +10,8 @@ import { RessursStatus } from '@navikt/familie-typer';
 import BarnIBrevSkjema from './BarnIBrev/BarnIBrevSkjema';
 import { barnIBrevÅrsakTilTittel, finnBarnIBrevÅrsak } from './barnIBrevÅrsak';
 import DeltBostedSkjema from './DeltBosted/DeltBostedSkjema';
-import { dokumentÅrsak, DokumentÅrsakPerson, useDokumentutsendingContext } from './DokumentutsendingContext';
+import { useDokumentutsendingContext } from './DokumentutsendingContext';
+import { dokumentÅrsak, DokumentÅrsakPerson } from './dokumentÅrsakTyper';
 import FritekstAvsnitt from './FritekstAvsnitt/FritekstAvsnitt';
 import KanSøkeSkjema from './KanSøke/KanSøkeSkjema';
 import { LeggTilBarnKnapp } from './LeggTilBarnKnapp';
