+++ src/frontend/sider/Fagsak/Dokumentutsending/dokumentÅrsakTyper.ts
@@ -1,0 +1,43 @@
+export enum DokumentÅrsakPerson {
+    DELT_BOSTED = 'DELT_BOSTED',
+    FØDSEL_MINDREÅRIG = 'FØDSEL_MINDREÅRIG',
+    FØDSEL_VERGEMÅL = 'FØDSEL_VERGEMÅL',
+    FØDSEL_GENERELL = 'FØDSEL_GENERELL',
+    KAN_SØKE = 'KAN_SØKE',
+    KAN_SØKE_EØS = 'KAN_SØKE_EØS',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER = 'TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_GJORT_VEDTAK_TIL_ANNEN_FORELDER = 'TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_GJORT_VEDTAK_TIL_ANNEN_FORELDER',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_ÅRLIG_KONTROLL = 'TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_ÅRLIG_KONTROLL',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER = 'TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER',
+    TIL_FORELDER_MED_SELVSTENDIG_RETT_VI_HAR_FÅTT_F016_KAN_SØKE_OM_BARNETRYGD = 'TIL_FORELDER_MED_SELVSTENDIG_RETT_VI_HAR_FÅTT_F016_KAN_SØKE_OM_BARNETRYGD',
+    KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV = 'KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV',
+    INNHENTE_OPPLYSNINGER_KLAGE = 'INNHENTE_OPPLYSNINGER_KLAGE',
+}
+
+export enum DokumentÅrsakInstitusjon {
+    INNHENTE_OPPLYSNINGER_KLAGE_INSTITUSJON = 'INNHENTE_OPPLYSNINGER_KLAGE_INSTITUSJON',
+}
+
+export type DokumentÅrsak = DokumentÅrsakPerson | DokumentÅrsakInstitusjon;
+
+export const dokumentÅrsak: Record<DokumentÅrsak, string> = {
+    DELT_BOSTED: 'Delt bosted',
+    FØDSEL_MINDREÅRIG: 'Fødsel mindreårig',
+    FØDSEL_VERGEMÅL: 'Fødsel vergemål',
+    FØDSEL_GENERELL: 'Fødsel generell',
+    KAN_SØKE: 'Kan søke',
+    KAN_SØKE_EØS: 'Kan søke EØS',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER:
+        'Til forelder omfattet norsk lovgivning - har fått en søknad fra annen forelder',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_GJORT_VEDTAK_TIL_ANNEN_FORELDER:
+        'Til forelder omfattet norsk lovgivning - har gjort vedtak til annen forelder',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_ÅRLIG_KONTROLL:
+        'Til forelder omfattet norsk lovgivning - varsel om årlig kontroll',
+    TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER:
+        'Til forelder omfattet norsk lovgivning - henter ikke registeropplysninger',
+    TIL_FORELDER_MED_SELVSTENDIG_RETT_VI_HAR_FÅTT_F016_KAN_SØKE_OM_BARNETRYGD:
+        'Til forelder med selvstendig rett vi har fått F016 - kan søke om barnetrygd',
+    KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV: 'Kan ha rett til pengestøtte fra Nav',
+    INNHENTE_OPPLYSNINGER_KLAGE: 'Innhente opplysninger klage',
+    INNHENTE_OPPLYSNINGER_KLAGE_INSTITUSJON: 'Innhente opplysninger klage',
+};
